+++ src/main/java/com/example/demo/domain/board/service/component/command/ViewIncreaseCommand.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain.board.service.component;
+package com.example.demo.domain.board.service.component.command;
 
 import com.example.demo.domain.board.Repository.BoardRepository;
 import com.example.demo.domain.board.Repository.ViewRepository;
